backend/requirements.txt
asgiref==3.4.1
async-timeout==4.0.3
boto3==1.24.91
botocore==1.27.91
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-crontab==0.7.1
django-decouple==2.1
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
gunicorn==20.1.0
hiredis==2.3.2
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.2
pandas==2.2.2
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
redis==5.0.3
requests==2.27.1
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
xmltodict==0.13.0
